pyproject.toml
[project]
name = "django-order-product"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2",
    "django-debug-toolbar>=5.1.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.14.1",
    "python-dotenv>=1.1.0",
]
